Split init script tests into 2 separate files
