+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/BaseInitScriptTest.groovy
@@ -19,6 +19,22 @@ import static nu.studer.teamcity.buildscan.agent.BuildScanServiceMessageInjector
 
 class BaseInitScriptTest extends Specification {
 
+    static final List<JdkCompatibleGradleVersion> NOT_SUPPORTED_GRADLE_VERSIONS = [
+            new JdkCompatibleGradleVersion(GradleVersion.version('3.5.1'), 7, 9),
+            new JdkCompatibleGradleVersion(GradleVersion.version('4.0.2'), 7, 9)
+    ]
+
+    static final List<JdkCompatibleGradleVersion> SUPPORTED_GRADLE_VERSIONS = [
+            new JdkCompatibleGradleVersion(GradleVersion.version('4.1'), 7, 9),
+            new JdkCompatibleGradleVersion(GradleVersion.version('4.10.3'), 7, 10),
+            new JdkCompatibleGradleVersion(GradleVersion.version('5.1.1'), 8, 11),
+            new JdkCompatibleGradleVersion(GradleVersion.version('5.6.4'), 8, 12),
+            new JdkCompatibleGradleVersion(GradleVersion.version('6.0.1'), 8, 13),
+            new JdkCompatibleGradleVersion(GradleVersion.version('6.7'), 8, 15),
+            new JdkCompatibleGradleVersion(GradleVersion.version('7.0.2'), 8, 16),
+            new JdkCompatibleGradleVersion(GradleVersion.version('7.4.2'), 8, 17),
+    ]
+
     static final String PUBLIC_BUILD_SCAN_ID = 'i2wepy2gr7ovw'
     static final String DEFAULT_SCAN_UPLOAD_TOKEN = 'scan-upload-token'
 
@@ -75,6 +91,9 @@ class BaseInitScriptTest extends Specification {
         initScriptFile = new File(testProjectDir, 'initscript.gradle')
         settingsFile = new File(testProjectDir, 'settings.gradle')
         buildFile = new File(testProjectDir, 'build.gradle')
+
+        settingsFile << ""
+        buildFile << ""
 
         FileUtil.copyResource(BuildScanServiceMessageInjector.class, '/' + BUILD_SCAN_INIT_GRADLE, initScriptFile)
     }
@@ -153,4 +172,38 @@ class BaseInitScriptTest extends Specification {
         out.toByteArray()
     }
 
+    static final class JdkCompatibleGradleVersion {
+
+        final GradleVersion gradleVersion
+        private final Integer jdkMin
+        private final Integer jdkMax
+
+        JdkCompatibleGradleVersion(GradleVersion gradleVersion, Integer jdkMin, Integer jdkMax) {
+            this.gradleVersion = gradleVersion
+            this.jdkMin = jdkMin
+            this.jdkMax = jdkMax
+        }
+
+        boolean isJvmVersionCompatible() {
+            def jvmVersion = getJvmVersion()
+            jdkMin <= jvmVersion && jvmVersion <= jdkMax
+        }
+
+        private static int getJvmVersion() {
+            String version = System.getProperty('java.version')
+            if (version.startsWith('1.')) {
+                Integer.parseInt(version.substring(2, 3))
+            } else {
+                Integer.parseInt(version.substring(0, version.indexOf('.')))
+            }
+        }
+
+        @Override
+        String toString() {
+            return "JdkCompatibleGradleVersion{" +
+                    "Gradle " + gradleVersion.version +
+                    ", JDK " + jdkMin + "-" + jdkMax +
+                    '}'
+        }
+    }
 }
